solves #49
When the buffer is changed we update the lint requests. Before we can
add new text to the lint requests we need to clear any requests refering
to lines that where changed since the last lint request update. For this
we have the `lint_req:clear_lines(start, stop)`. It however removed lint
requests from front to back causing. Popping the lint requests from the
front shifted the following requests meaning the wrong requests where
deleted. I fixed this.

This also adds a runtime check if the lint requests do not contain
duplicates. This should be hidden behind a debug flag if it proves to
harm performance.