Merge dev branch (#54)
* adds infrastructure for preprosessing

* lintreq are now buffer specific

* fixes text extraction and underline location

* refactors buffer specific config/objects

* changes config format

* updates readme to reflect config changes

* latex url/domain/ filtering added

* refactor latex preprocessing

* fix col and refactor pattern preprocessor

* adds support for query directives

* fixes #47

* adds note on latex file detection behaviour

* Refactors issues into IssueList and Issue object

* adds hover menu for fixing spell errors on the go

* fixes float window size

* slight improvements to auto install

* fixes vale style install

* Hover menu replace implemented using lsp utils text edit

* introduces IssueList and fixes #48

* fixes error with overlapping extmarks

* fixes issue not relinked to new id

* workaround for the system linting popups (I think)

* verious small fixes

* Fix extmark color adjustment changing extmark id

* Update existing marks color instead of removing the mark and inserting
another

* adds a lot of debug statements to illustrate bug

* solves #49

When the buffer is changed we update the lint requests. Before we can
add new text to the lint requests we need to clear any requests refering
to lines that where changed since the last lint request update. For this
we have the `lint_req:clear_lines(start, stop)`. It however removed lint
requests from front to back causing. Popping the lint requests from the
front shifted the following requests meaning the wrong requests where
deleted. I fixed this.

This also adds a runtime check if the lint requests do not contain
duplicates. This should be hidden behind a debug flag if it proves to
harm performance.

* fixes #51